requirements.txt
langchain>=0.3.14
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.7.0
PyMuPDF>=1.23.0
requests>=2.31.0
ragas>=0.2.0
chainlit>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pydantic>=2.0.0
rank-bm25>=0.2.2
